feat: jump to recent record on click
