+++ insights/api/__init__.py
@@ -563,66 +563,3 @@ def contact_team(message_type, message_content, is_critical=False):
     except Exception as e:
         frappe.log_error(e)
         frappe.throw("Something went wrong. Please try again later.")
-
-
-@frappe.whitelist()
-def create_last_viewed_log(record_type, record_name):
-    recordToDoctype = {
-        "Query": "Insights Query",
-        "Dashboard": "Insights Dashboard",
-        "NotebookPage": "Insights Notebook Page",
-    }
-    try:
-        comment = frappe.get_doc(
-            {
-                "doctype": "Comment",
-                "comment_type": "Comment",
-                "comment_email": frappe.session.user,
-                "reference_doctype": recordToDoctype.get(record_type),
-                "reference_name": record_name,
-                "content": "Last Viewed",
-            }
-        )
-        comment.insert(ignore_permissions=True)
-    except BaseException:
-        pass
-
-
-@frappe.whitelist()
-def get_last_viewed_records():
-    Comment = frappe.qb.DocType("Comment")
-    TRACKED_DOCTYPES = ["Insights Query", "Insights Dashboard", "Insights Notebook Page"]
-    records = (
-        frappe.qb.from_(Comment)
-        .select(Comment.reference_doctype, Comment.reference_name, Comment.creation)
-        .where(
-            (Comment.comment_email == frappe.session.user)
-            & Comment.reference_doctype.isin(TRACKED_DOCTYPES)
-            & (Comment.comment_type == "Comment")
-            & (Comment.content == "Last Viewed")
-        )
-        .orderby(Comment.creation, order=frappe.qb.desc)
-        .groupby(Comment.reference_doctype, Comment.reference_name)
-        .limit(10)
-        .run(as_dict=True)
-    )
-
-    # fetch titles for each record
-    # we run a separate query for each doctype to avoid a large IN query
-    docnames_by_doctype = {}
-    for record in records:
-        docnames_by_doctype.setdefault(record.reference_doctype, []).append(record.reference_name)
-
-    for doctype, docnames in docnames_by_doctype.items():
-        titles = frappe.get_all(
-            doctype,
-            filters={"name": ["in", docnames]},
-            fields=["name", "title"],
-        )
-        for title in titles:
-            for record in records:
-                if record.reference_doctype == doctype and record.reference_name == title.name:
-                    record["title"] = title.title
-                    break
-
-    return records
